// src/api/models.ts
export const ApiTargets = {
  BOOTSTRAP: '/accounts/bootstrap',
  SEARCH: '/search/query',
  QTREE: '/search/qtree',
  BIGQUERY: '/search/bigquery',
  EXPORT: '/export',
  SERVICE_AUTHOR_NETWORK: '/vis/author-network',
  SERVICE_PAPER_NETWORK: '/vis/paper-network',
  SERVICE_WORDCLOUD: '/vis/word-cloud',
  SERVICE_METRICS: '/metrics',
  SERVICE_OBJECTS: '/objects',
  SERVICE_OBJECTS_QUERY: '/objects/query',
  SERVICE_CITATION_HELPER: '/citation_helper',
  SERVICE_AUTHOR_AFFILIATION_EXPORT: '/authoraff',
  MYADS_STORAGE: '/vault',
  MYADS_STORAGE_QUERY: '/vault/query',
  MYADS_NOTIFICATIONS: '/vault/notifications',
  MYADS_NOTIFICATIONS_QUERY: '/vault/notification_query',
  LINK_SERVERS: '/vault/configuration/link_servers',
  AUTHOR_AFFILIATION_SEARCH: '/author-affiliation/search',
  AUTHOR_AFFILIATION_EXPORT: '/author-affiliation/export',
  RESOLVER: '/resolver',
  CSRF: '/accounts/csrf',
  USER: '/accounts/user',
  USER_DATA: '/vault/user-data',
  SITE_CONFIGURATION: '/vault/configuration',
  TOKEN: '/accounts/token',
  INFO: '/accounts/info',
  LOGIN: '/accounts/login',
  LOGOUT: '/accounts/logout',
  REGISTER: '/accounts/register',
  VERIFY: '/accounts/verify',
  DELETE: '/accounts/delete',
  RESET_PASSWORD: '/accounts/reset-password',
  CHANGE_PASSWORD: '/accounts/change-password',
  CHANGE_EMAIL: '/accounts/change-email',
  RECOMMENDER: '/recommender',
  GRAPHICS: '/graphics',
  FEEDBACK: '/feedback/userfeedback',
  LIBRARY_IMPORT_CLASSIC_AUTH: '/harbour/auth/classic',
  LIBRARY_IMPORT_CLASSIC_MIRRORS: '/harbour/mirrors',
  LIBRARY_IMPORT_CLASSIC_TO_BBB: '/biblib/classic',
  LIBRARY_IMPORT_ADS2_AUTH: '/harbour/auth/twopointoh',
  LIBRARY_IMPORT_ADS2_TO_BBB: '/biblib/twopointoh',
  LIBRARY_IMPORT_ZOTERO: '/harbour/export/twopointoh/zotero',
  LIBRARY_IMPORT_MENDELEY: '/harbour/export/twopointoh/mendeley',
  LIBRARY_IMPORT_CREDENTIALS: '/harbour/user',
  ORCID_PREFERENCES: '/orcid/preferences',
  ORCID: '/orcid',
  ORCID_NAME: '/orcid/orcid-name',
  ORCID_WORKS: 'orcid-works',
  ORCID_WORK: 'orcid-work',
  ORCID_PROFILE: 'orcid-profile',
  ORCID_EXCHANGE_TOKEN: '/orcid/exchangeOAuthCode',
  LIBRARIES: '/biblib/libraries',
  LIBRARY_TRANSFER: '/biblib/transfer',
  LIBRARY_OPERATION: 'biblib/libraries/operations',
  LIBRARY_QUERY: 'biblib/query',
  LIBRARY_NOTES: 'biblib/notes',
  DOCUMENTS: '/biblib/documents',
  PERMISSIONS: '/biblib/permissions',
  REFERENCE: '/reference/text',
} as const;